requirements.txt
africastalking==1.2.7
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.2
contextlib2==21.6.0
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
idna==3.4
Pillow==10.1.0
PyJWT==1.7.1
python-decouple==3.8
pytz==2023.3.post1
requests==2.31.0
schema==0.7.5
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.1.0
